pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qugeister"
version = "1.0.0"
description = "Quantum Geister AI Competition System"
authors = [
    {name = "Qugeister Development Team"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "Intended Audience :: Science/Research",
    "Topic :: Games/Entertainment :: Board Games",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=1.11.0",
    "pennylane>=0.28.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "PyYAML>=6.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0"
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0"
]

[project.scripts]
qugeister = "qugeister.cli.main:main"

[project.urls]
Homepage = "https://github.com/qugeister/qugeister"
Repository = "https://github.com/qugeister/qugeister"
Documentation = "https://qugeister.readthedocs.io/"
"Bug Reports" = "https://github.com/qugeister/qugeister/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
qugeister = [
    "web/templates/*.html",
    "web/static/css/*.css",
    "web/static/js/*.js", 
    "web/static/images/*"
]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

# MyPy configuration  
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/qugeister --cov-report=html --cov-report=term-missing"